Refactor the select classes
