+++ doc/Controls.js.html
@@ -307,6 +307,35 @@ export function initEditMode(editMode) {
         editMode.init();
     });
 }
+
+/**
+ * Set listeners on the buttons to change selection modes.
+ */
+export function initSelectionButtons() {
+    $("#selByNeume").on("click", function(){
+        if (!$("#selByNeume").hasClass("is-active")){
+            $("#selByNeume").addClass("is-active");
+            $("#selByNc").removeClass("is-active");
+            $("#selByStaff").removeClass("is-active");
+        }           
+    });
+
+    $("#selByNc").on("click", function(){
+        if (!$("#selByNc").hasClass("is-active")) {
+            $("#selByNc").addClass("is-active");
+            $("#selByNeume").removeClass("is-active");
+            $("#selByStaff").removeClass("is-active");
+        }
+    });
+
+    $("#selByStaff").on("click", function () {
+        if (!$("#selByStaff").hasClass("is-active")) {
+            $("#selByStaff").addClass("is-active");
+            $("#selByNc").removeClass("is-active");
+            $("#selByNeume").removeClass("is-active");
+        }
+    });
+}
 </code></pre>
         </article>
     </section>
@@ -317,13 +346,13 @@ export function initEditMode(editMode) {
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Select.html">Select</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Select.ClickSelect.html">ClickSelect</a></li><li><a href="module-Select.DragSelect.html">DragSelect</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 14:27:49 GMT-0400 (EDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 14:46:39 GMT-0400 (EDT)
 </footer>
 
 <script> prettyPrint(); </script>
