+++ doc/index.html
@@ -42,6 +42,37 @@
 
 
 
+    <section>
+        <article><h1>Neon3</h1><p><a href="https://travis-ci.org/DDMAL/Neon2"><img src="https://travis-ci.org/DDMAL/Neon2.svg?branch=develop" alt="Build_Status"></a>
+<a href="https://opensource.org/licenses/MIT"><img src="https://img.shields.io/badge/License-MIT-yellow.svg" alt="License: MIT"></a></p>
+<p><strong>N</strong>eume <strong>E</strong>ditor <strong>ON</strong>line.</p>
+<p>Neon3 is a browser-based music notation editor written in JavaScript using the Verovio music engraving library. The editor can be used to manipulate digitally encoded early musical scores in square-note notation.</p>
+<p>Neon2 is a re-write of <a href="https://github.com/DDMAL/Neon.js">Neon.JS</a> using a modified version of <a href="https://github.com/DDMAL/verovio">Verovio</a> to render MEI-Neume files according to the MEI 4.0 specifications. </p>
+<h2>Requirements</h2><ul>
+<li><a href="https://yarnpkg.com/en/docs/install">yarn</a>: <ul>
+<li><code>brew install yarn</code> on Mac</li>
+</ul>
+</li>
+</ul>
+<h2>Setup</h2><ol>
+<li><p>Install the dependencies using yarn:</p>
+<pre class="prettyprint source"><code>yarn install</code></pre></li>
+<li><p>Build webpack:</p>
+<pre class="prettyprint source"><code>yarn build</code></pre></li>
+<li><p>Start the server:</p>
+<pre class="prettyprint source"><code>yarn start</code></pre></li>
+<li><p>Access the page at: <a href="http://localhost:8080">http://localhost:8080</a>.</p>
+</li>
+</ol>
+<h2>Instructions</h2><p>Neon.js has two main modes: viewer and editor. To learn how to use both, <a href="https://github.com/DDMAL/Neon2/wiki/Instructions">read the instructions on our wiki.</a></p>
+<h2>Test</h2><p>Follow the instructions from above first. The tests for Neon2 use <a href="https://docs.seleniumhq.org/">Selenium</a> and so require a web browser (<a href="https://mozilla.org/firefox">Firefox</a>) and its driver (<a href="https://github.com/mozilla/geckodriver">geckodriver</a>).
+On Mac install these with Homebrew:</p>
+<pre class="prettyprint source"><code>brew cask install firefox
+brew install geckodriver</code></pre><p>Then you can run the tests locally using <code>yarn test</code>. We use <a href="https://facebook.github.io/jest/">jest</a> to script our tests.</p>
+<p><em>These tests require the server to be running on <code>localhost:8080</code></em></p>
+<h2>Verovio</h2><p>Verovio is present as an npm package under <code>src/verovio-dev</code> with the name <code>verovio-dev</code>. Its contents come from the <code>emscripten/npm-dev</code> folder in a Verovio project folder.</p></article>
+    </section>
+
 
 
 
@@ -50,13 +81,13 @@
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Select.html">Select</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Select.ClickSelect.html">ClickSelect</a></li><li><a href="module-Select.DragSelect.html">DragSelect</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 14:27:49 GMT-0400 (EDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 14:46:39 GMT-0400 (EDT)
 </footer>
 
 <script> prettyPrint(); </script>
