+++ doc/module-Select.DragSelect.html
@@ -28,7 +28,8 @@
 
 <header>
     
-        <h2><span class="attribs"><span class="type-signature"></span></span>DragSelect<span class="signature">(dragHandler, zoomHandler, groupingHandler, neonView)</span><span class="type-signature"></span></h2>
+        <h2><span class="attribs"><span class="type-signature"></span></span>
+            <span class="ancestors"><a href="module-Select.html">Select</a>.</span>DragSelect<span class="signature">(dragHandler, zoomHandler, groupingHandler, neonView)</span><span class="type-signature"></span></h2>
         
     
 </header>
@@ -211,7 +212,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="DragSelect.js.html">DragSelect.js</a>, <a href="DragSelect.js.html#line12">line 12</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line222">line 222</a>
     </li></ul></dd>
     
 
@@ -389,7 +390,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="DragSelect.js.html">DragSelect.js</a>, <a href="DragSelect.js.html#line165">line 165</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line375">line 375</a>
     </li></ul></dd>
     
 
@@ -475,7 +476,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="DragSelect.js.html">DragSelect.js</a>, <a href="DragSelect.js.html#line58">line 58</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line268">line 268</a>
     </li></ul></dd>
     
 
@@ -561,7 +562,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="DragSelect.js.html">DragSelect.js</a>, <a href="DragSelect.js.html#line79">line 79</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line289">line 289</a>
     </li></ul></dd>
     
 
@@ -647,94 +648,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="DragSelect.js.html">DragSelect.js</a>, <a href="DragSelect.js.html#line30">line 30</a>
-    </li></ul></dd>
-    
-
-    
-
-    
-
-    
-</dl>
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-        
-            
-
-    
-
-    
-    <h4 class="name" id="~unselect"><span class="type-signature">(inner) </span>unselect<span class="signature">()</span><span class="type-signature"></span></h4>
-    
-
-    
-
-
-
-<div class="description">
-    Unselect all selected elements and run undo any extra
-actions.
-</div>
-
-
-
-
-
-
-
-
-
-
-
-
-
-<dl class="details">
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-    <dt class="tag-source">Source:</dt>
-    <dd class="tag-source"><ul class="dummy"><li>
-        <a href="DragSelect.js.html">DragSelect.js</a>, <a href="DragSelect.js.html#line196">line 196</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line240">line 240</a>
     </li></ul></dd>
     
 
@@ -938,7 +852,7 @@ actions.
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="DragSelect.js.html">DragSelect.js</a>, <a href="DragSelect.js.html#line184">line 184</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line394">line 394</a>
     </li></ul></dd>
     
 
@@ -982,13 +896,13 @@ actions.
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="DragSelect.html">DragSelect</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li><li><a href="Select.html">Select</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-Color.html">Color</a></li><li><a href="module-Contents.html">Contents</a></li><li><a href="module-Controls.html">Controls</a></li><li><a href="module-Cursor.html">Cursor</a></li><li><a href="module-Neon.html">Neon</a></li><li><a href="module-Select.html">Select</a></li><li><a href="module-Zoom.html">Zoom</a></li></ul><h3>Classes</h3><ul><li><a href="DragHandler.html">DragHandler</a></li><li><a href="EditMode.html">EditMode</a></li><li><a href="GroupingHandler.html">GroupingHandler</a></li><li><a href="InfoBox.html">InfoBox</a></li><li><a href="InsertHandler.html">InsertHandler</a></li><li><a href="module-Neon-Neon.html">Neon</a></li><li><a href="module-Select.ClickSelect.html">ClickSelect</a></li><li><a href="module-Select.DragSelect.html">DragSelect</a></li><li><a href="module-Zoom.ViewBox.html">ViewBox</a></li><li><a href="module-Zoom-ZoomHandler.html">ZoomHandler</a></li><li><a href="NeonView.html">NeonView</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 14:27:49 GMT-0400 (EDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Thu Jul 19 2018 14:46:39 GMT-0400 (EDT)
 </footer>
 
 <script> prettyPrint(); </script>
