+++ doc/module-Select.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title>JSDoc: Class: DragHandler</title>
+    <title>JSDoc: Module: Select</title>
 
     <script src="scripts/prettify/prettify.js"> </script>
     <script src="scripts/prettify/lang-css.js"> </script>
@@ -17,7 +17,7 @@
 
 <div id="main">
 
-    <h1 class="page-title">Class: DragHandler</h1>
+    <h1 class="page-title">Module: Select</h1>
 
     
 
@@ -28,7 +28,6 @@
 
 <header>
     
-        <h2><span class="attribs"><span class="type-signature"></span></span>DragHandler<span class="signature">(neonView)</span><span class="type-signature"></span></h2>
         
     
 </header>
@@ -38,138 +37,7 @@
     
         
 
-    
-
-    
-    <h4 class="name" id="DragHandler"><span class="type-signature"></span>new DragHandler<span class="signature">(neonView)</span><span class="type-signature"></span></h4>
-    
-
-    
-
-
-
-<div class="description">
-    Handle the dragging of musical elements and communicate actions.
-</div>
-
-
-
-
-
-
-
-
-
-    <h5>Parameters:</h5>
-    
-
-<table class="params">
-    <thead>
-    <tr>
         
-        <th>Name</th>
-        
-
-        <th>Type</th>
-
-        
-
-        
-
-        <th class="last">Description</th>
-    </tr>
-    </thead>
-
-    <tbody>
-    
-
-        <tr>
-            
-                <td class="name"><code>neonView</code></td>
-            
-
-            <td class="type">
-            
-                
-<span class="param-type"><a href="NeonView.html">NeonView</a></span>
-
-
-            
-            </td>
-
-            
-
-            
-
-            <td class="description last">The NeonView parent object.</td>
-        </tr>
-
-    
-    </tbody>
-</table>
-
-
-
-
-
-
-<dl class="details">
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-    <dt class="tag-source">Source:</dt>
-    <dd class="tag-source"><ul class="dummy"><li>
-        <a href="DragHandler.js.html">DragHandler.js</a>, <a href="DragHandler.js.html#line6">line 6</a>
-    </li></ul></dd>
-    
-
-    
-
-    
-
-    
-</dl>
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
     
     </div>
 
@@ -177,6 +45,16 @@
 
     
 
+    
+        <h3 class="subsection-title">Classes</h3>
+
+        <dl>
+            <dt><a href="module-Select.ClickSelect.html">ClickSelect</a></dt>
+            <dd></dd>
+        
+            <dt><a href="module-Select.DragSelect.html">DragSelect</a></dt>
+            <dd></dd>
+        </dl>
     
 
     
@@ -196,7 +74,7 @@
     
 
     
-    <h4 class="name" id="~dragInit"><span class="type-signature">(inner) </span>dragInit<span class="signature">()</span><span class="type-signature"></span></h4>
+    <h4 class="name" id="~unselect"><span class="type-signature">(inner) </span>unselect<span class="signature">()</span><span class="type-signature"></span></h4>
     
 
     
@@ -204,7 +82,8 @@
 
 
 <div class="description">
-    Initialize the dragging action and handler for selected elements.
+    Unselect all selected elements and run undo any extra
+actions.
 </div>
 
 
@@ -248,7 +127,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="DragHandler.js.html">DragHandler.js</a>, <a href="DragHandler.js.html#line13">line 13</a>
+        <a href="Select.js.html">Select.js</a>, <a href="Select.js.html#line407">line 407</a>
     </li></ul></dd>
